==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "/cloud/project/R/01_load.R")
source(file = "/cloud/project/R/02_clean.R")
source(file = "/cloud/project/R/03_augment.R")
source(file = "/cloud/project/R/04_summary_models.R")
source(file = "/cloud/project/R/05_binomial_regression.R")
source(file = "/cloud/project/R/06_cholestoral_analysis.R")
source(file = "/cloud/project/R/07_explorative_data_an.R")
source(file = "/cloud/project/R/08_unsupervised_analysis.R")
rmarkdown::render("/cloud/project/doc/Presentation_IOslides.Rmd")
browseURL("/cloud/project/doc/Presentation_IOslides.html")
